% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_reexports.R
\docType{data}
\name{.ggplot2}
\alias{.ggplot2}
\title{placeholder for ggplot2 namespace backup}
\format{An object of class \code{environment} of length 854.}
\usage{
.ggplot2
}
\description{
placeholder for ggplot2 namespace backup
}
\keyword{datasets}
