% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_reexports.R
\name{.onLoad}
\alias{.onLoad}
\title{cache ggplot2 objects into ggtheme, allowing them to be modified directly
within ggplot2 and restored as needed}
\usage{
.onLoad(libname, pkgname)
}
\description{
This is necessary because ggplot2's internal versions of scale_* functions
will be used with precedence over user functions or masked versions available
in a global environment.
}
