% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_reexports.R
\name{modify_ggplot2_binding}
\alias{modify_ggplot2_binding}
\title{modify a locked binding within an environment (ggplot2 & ggtheme by default)}
\usage{
modify_ggplot2_binding(x, value, ...)
}
\description{
modify a locked binding within an environment (ggplot2 & ggtheme by default)
}
