% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_reexports.R
\name{restore_ggplot2_binding}
\alias{restore_ggplot2_binding}
\title{restore a single binding within ggplot2 namespace back to original value}
\usage{
restore_ggplot2_binding(x = getNamespaceExports("ggplot2"), ...)
}
\description{
restore a single binding within ggplot2 namespace back to original value
}
