% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_pub.R
\name{scale_colour_pub}
\alias{scale_colour_pub}
\title{The color palette is also designed with the help of color brewer using bold and contrasting colors so, one can easily distinguish any two colors .}
\usage{
scale_colour_pub(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
The color palette is also designed with the help of color brewer using bold and contrasting colors so, one can easily distinguish any two colors .
}
\examples{
gglm(group = "Species", label.y = 0.99) + ggtheme::theme_pub() +
 ggtheme::scale_colour_pub()
}
\author{
Koundinya Desiraju
}
