% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_pub.R
\name{scale_fill_pub}
\alias{scale_fill_pub}
\title{The color palette is also designed with the help of color brewer using bold and contrasting colors so, one can easily distinguish any two colors .}
\usage{
scale_fill_pub(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
The color palette is also designed with the help of color brewer using bold and contrasting colors so, one can easily distinguish any two colors .
}
\examples{
ggplot(mtcars, aes(factor(carb),fill=factor(carb))) + geom_bar(alpha=0.7) + ggtheme::theme_pub() +
 ggtheme::scale_fill_pub()
}
\author{
Koundinya Desiraju
}
