% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_reexports.R
\name{search_ggplot2_env}
\alias{search_ggplot2_env}
\title{get ggplot2 environment attached in global search path}
\usage{
search_ggplot2_env()
}
\description{
get ggplot2 environment attached in global search path
}
