% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme.R
\name{set_theme_value}
\alias{set_theme_value}
\title{Title}
\usage{
set_theme_value(..., set = NULL, theme, quoted = FALSE)
}
\arguments{
\item{...}{}

\item{set}{}

\item{theme}{}
}
\value{

}
\description{
Title
}
