% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme_general.R
\name{simpletheme}
\alias{simpletheme}
\title{Title}
\usage{
simpletheme(
  text_size = 8,
  text_color = "black",
  line_color = "black",
  line_size = 0.4672897,
  axis.text.x.angle = 0,
  axis.text.x.hjust = 0,
  axis.text.x.vjust = 0,
  ...
)
}
\arguments{
\item{text_size}{The text size}

\item{text_color}{Text color for axis. Default: black.}

\item{line_color}{Text color for axis. Default: black.}

\item{line_size}{The line size}

\item{axis.text.x.angle}{The angle of text in x axis.}

\item{axis.text.x.vjust}{}

\item{...}{others parameters.}
}
\value{

}
\description{
Title
}
