% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_grid.R
\name{theme_grid}
\alias{theme_grid}
\title{Modify the grid of ggplot2}
\usage{
theme_grid(colour = "gray90", linetype = 1)
}
\arguments{
\item{colour}{grid line colour}

\item{linetype}{grid line type, one of 0, 1, 2, 3}
}
\value{

}
\description{
Modify the grid of ggplot2
}
\examples{
theme_grid(colour = "gray90", linetype = 0)
}
\author{
ZhonghuiGai
}
