% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_prism.R
\name{theme_prism}
\alias{theme_prism}
\title{GraphPad Prism themes}
\usage{
theme_prism(base_size = 16, base_family = "serif", angle = 45, border = FALSE)
}
\arguments{
\item{base_size}{`numeric`. Base font size, given in `"pt"`.}

\item{base_family}{`string`. Base font family, default is `"Times New Roman"`.}

\item{angle}{`integer`. Angle of x_axis text in degrees.
One of: `-30, 45, 90, 270`.}

\item{border}{`logical`. Should a border be drawn around the plot?
Clipping will occur unless e.g. `coord_cartesian(clip = "off")` is used.}
}
\value{
Returns a list-like object of class _theme_.
}
\description{
GraphPad Prism themes
}
