% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_size.R
\name{theme_size}
\alias{theme_size}
\title{Change the size of the ggplot}
\usage{
theme_size(size = 14)
}
\arguments{
\item{size}{a numeric, the default value is 14}
}
\value{

}
\description{
Change the size of the ggplot
}
\examples{
theme_size(size = 14)
}
\author{
Zhonghui Gai
}
