% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pub.R
\name{theme_pub}
\alias{theme_pub}
\title{ggplot theme for publication ready Plots}
\usage{
theme_pub(base_size = 14, base_family = "sans")
}
\arguments{
\item{base_size}{the default value is 14}

\item{base_family}{the default value is "sans"}
}
\value{

}
\description{
ggplot theme for publication ready Plots
}
\examples{
gglm::gglm() + ggtheme::theme_pub()
gglm(group = "Species", label.y = 0.99) + ggtheme::theme_pub() +
ggprism::scale_color_prism(palette = "floral")
}
\author{
Koundinya Desiraju
}
